// src/modules/user/user.types.ts
import type { Document, Types } from "mongoose";
import type { UserRole, UserStatus } from "../../config/constants";

export interface IUser {
  name: string;
  email: string;
  password: string;
  role: UserRole;
  status: UserStatus;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUserDocument extends IUser, Document {
  _id: Types.ObjectId;
  comparePassword(candidatePassword: string): Promise<boolean>;
}